requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow-cpu==2.15.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
requests==2.31.0
pandas==2.0.3